cull out most of the pubs, modularizes supercond
